fix: validate associated users

+++ backend/src/main/java/vn/tailinh/internmatching/service/RoleService.java
@@ -1,6 +1,7 @@
 package vn.tailinh.internmatching.service;
 
 import lombok.RequiredArgsConstructor;
+import vn.tailinh.internmatching.controller.RoleController;
 import vn.tailinh.internmatching.entity.Permission;
 import vn.tailinh.internmatching.dto.response.ResultPaginationResponse;
 import vn.tailinh.internmatching.exception.IdInvalidException;
@@ -22,9 +23,15 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 public class RoleService {
+  private final RoleController roleController;
   private final RoleRepository roleRepository;
   private final PermissionRepository permissionRepository;
 
+
+
+
+
+  
   public Role create(Role role) throws Exception {
     if (this.roleRepository.existsByName(
         role.getName())) {
@@ -41,6 +48,8 @@ public class RoleService {
     return this.roleRepository.save(role);
   }
 
+
+
   public Role fetchRoleById(Long id) throws Exception {
     Optional<Role> role = this.roleRepository.findById(id);
     if (role.isPresent()) {
@@ -49,6 +58,9 @@ public class RoleService {
       throw new IdInvalidException("The specified Role ID is invalid");
     }
   }
+
+
+
 
   public Role fetchRoleByName(String name) {
     return this.roleRepository.findByName(name);
@@ -73,11 +85,18 @@ public class RoleService {
     return this.roleRepository.save(currentRole);
   }
 
-  public void delete(Long id) throws Exception {
+
+
+  public void delete(Long id ) throws Exception {
     Role role = this.fetchRoleById(id);
 
+    if(role.getUsers() != null && !role.getUsers().isEmpty()) {
+      throw new IdInvalidException("Cannot delete this role because there are users associated with it");
+    }
     this.roleRepository.delete(role);
   }
+
+
 
   public ResultPaginationResponse fetchAll(Specification<Role> spec, Pageable pageable) {
     Page<Role> rolePage = this.roleRepository.findAll(spec, pageable);
